+++ sickbeard/search.py
@@ -457,7 +457,7 @@ def search_for_needed_episodes(episodes):
     threading.currentThread().name = orig_thread_name
 
     if not len(providers):
-        logger.log('No NZB/Torrent sources enabled in Search Provider options to do recent searches', logger.WARNING)
+        logger.log('No NZB/Torrent providers in Media Providers/Options are enabled to match recent episodes', logger.WARNING)
     elif not search_done:
         logger.log('Failed recent search of %s enabled provider%s. More info in debug log.' % (
             len(providers), helpers.maybe_plural(len(providers))), logger.ERROR)
@@ -465,7 +465,7 @@ def search_for_needed_episodes(episodes):
     return found_results.values()
 
 
-def search_providers(show, episodes, manual_search=False, torrent_only=False, try_other_searches=False):
+def search_providers(show, episodes, manual_search=False, torrent_only=False, try_other_searches=False, scheduled=False):
     found_results = {}
     final_results = []
 
@@ -474,7 +474,8 @@ def search_providers(show, episodes, manual_search=False, torrent_only=False, tr
     orig_thread_name = threading.currentThread().name
 
     provider_list = [x for x in sickbeard.providers.sortedProviderList() if x.is_active() and x.enable_backlog and
-                     (not torrent_only or x.providerType == GenericProvider.TORRENT)]
+                     (not torrent_only or x.providerType == GenericProvider.TORRENT) and
+                     (not scheduled or x.enable_scheduled_backlog)]
     for cur_provider in provider_list:
         if cur_provider.anime_only and not show.is_anime:
             logger.log(u'%s is not an anime, skipping' % show.name, logger.DEBUG)
@@ -773,9 +774,9 @@ def search_providers(show, episodes, manual_search=False, torrent_only=False, tr
             break
 
     if not len(provider_list):
-        logger.log('No NZB/Torrent sources enabled in Search Provider options to do backlog searches', logger.WARNING)
+        logger.log('No NZB/Torrent providers in Media Providers/Options are allowed for active searching', logger.WARNING)
     elif not search_done:
-        logger.log('Failed backlog search of %s enabled provider%s. More info in debug log.' % (
+        logger.log('Failed active search of %s enabled provider%s. More info in debug log.' % (
             len(provider_list), helpers.maybe_plural(len(provider_list))), logger.ERROR)
 
     return final_results