+++ sickbeard/search_queue.py
@@ -339,7 +339,7 @@ class RecentSearchQueueItem(generic_queue.QueueItem):
             threads[-1].start()
 
         if not len(providers):
-            logger.log('No NZB/Torrent sources enabled in Search Provider options for cache update', logger.WARNING)
+            logger.log('No NZB/Torrent providers in Media Providers/Options are enabled to match recent episodes', logger.WARNING)
 
         if threads:
             # wait for all threads to finish
@@ -429,7 +429,8 @@ class BacklogQueueItem(generic_queue.QueueItem):
             logger.log(u'Beginning backlog search for: [%s]' % self.show.name)
             search_result = search.search_providers(
                 self.show, self.segment, False,
-                try_other_searches=(not self.standard_backlog or not self.limited_backlog))
+                try_other_searches=(not self.standard_backlog or not self.limited_backlog),
+                scheduled=self.standard_backlog)
 
             if search_result:
                 for result in search_result: